refactor(section): Section 서비스와 컨트롤러가 Line에 포함되도록 변경
Section은 Line의 Cascade에 의해 관리되므로 Line에서 관리하는게 자연스럽다
별도의 Section 서비스와 컨트롤러가 존재할 경우 Section의 영속성을 직접 관리
하는것처럼 보일 수 있다.
