+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -57,4 +57,12 @@ public class LineController {
         lineService.deleteLine(id);
         return ResponseEntity.noContent().build();
     }
+
+    @PostMapping(value = "/{id}/sections")
+    public ResponseEntity<LineResponse> addSection(
+            @PathVariable Long id,
+            @RequestBody SectionRequest sectionRequest) {
+        LineResponse line = lineService.saveSection(id, sectionRequest);
+        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
+    }
 }
